Eliminate TOCTOU races, add symlink protection, and comprehensive hardening
Session subsystem overhaul:
- Hold exclusive lock across full read-modify-write cycle (TOCTOU fix)
- Add TrackedPid struct with start_time for PID reuse detection
- Add fsync after writes for crash safety, seek(0) for defensive coding
- Count and report kill failures instead of silently ignoring
- Backward-compatible deserialization for existing sessions.json

Security hardening:
- Add symlink protection for all predictable file write paths
- Verify daemon PID identity before sending SIGTERM in stop()
- Set ~/.proc-janitor/ directory permissions to 0o700
- Allow /var/log/ in safe log path check while blocking other /var/ paths

Performance and correctness:
- Single-process refresh in verify_process_identity (was full table scan)
- Browser child process wait() to prevent zombies
- Cycle detection in has_target_descendant and print_subtree
- Append mode for daemon.out/err to preserve logs across restarts
- BufReader for print_recent_logs to avoid loading entire file

Code quality:
- Extract use_color() to shared util.rs (was duplicated in 4 files)
- Add empty targets warning in scanner and doctor health check
- Merge format_toml_array/whitelist into parameterized function
- Add scan_interval validation in Config::validate()
- Support editor flags in EDITOR env var (e.g. "code --wait")
- Fix integer casts, clippy warnings, integration test JSON parsing
- Remove dead code (cleanup_stale, daemon wake/stop stubs)
- Document container PID 1 limitation on is_orphan()

+++ src/cli.rs
@@ -72,8 +72,8 @@ pub enum Commands {
         follow: bool,
 
         /// Number of lines to show
-        #[arg(short = 'n', long, default_value = "50")]
-        lines: usize,
+        #[arg(short = 'n', long, default_value = "50", value_parser = clap::value_parser!(u64).range(1..))]
+        lines: u64,
     },
 
     /// Session-based process tracking
